refactor(#289): extract migrations from db/schema.ts
Move all 12 migration functions into a dedicated migrations.ts file.
schema.ts now contains only the SCHEMA constant (~181 lines) plus
re-exports for backward compatibility.
